Added components for billing and preferences
